Refactor: transmit trust to wallet model
