+++ server/models/Wallet.js
@@ -1,4 +1,5 @@
 const WalletRepository = require("../repositories/WalletRepository");
+const TrustRepository = require("../repositories/TrustRepository");
 const HttpError = require("../utils/HttpError");
 const Crypto = require('crypto');
 const expect = require("expect-runtime");
@@ -10,6 +11,7 @@ class Wallet{
     expect(id).number();
     this.id = id;
     this.walletRepository = new WalletRepository();
+    this.trustRepository = new TrustRepository();
   }
 
   async authorize(password){
@@ -28,8 +30,33 @@ class Wallet{
     }
   }
 
+  async getTrustRelationships(){
+    return await this.trustRepository.get();
+  }
+
   async toJSON(){
     return await this.walletRepository.getById(this.id);
+  }
+
+  async request(requestType, walletName){
+    expect(requestType, () => new HttpError(400, `The trust request type must be one of ${Object.keys(TrustRepository.ENTITY_TRUST_REQUEST_TYPE).join(',')}`))
+      .oneOf(Object.keys(TrustRepository.ENTITY_TRUST_REQUEST_TYPE));
+    expect(walletName, () => new HttpError(400, "Invalid wallet name"))
+      .match(/\S+/);
+    
+    //get wallet id
+    const wallet = await this.walletRepository.getByName((walletName));
+    const result = await this.trustRepository.create({
+      request_type: requestType,
+      target_entity_id: wallet.id,
+    });
+    return result;
+  }
+  
+  async accept(trustRelationshipId){
+    const trustRelationship = await this.trustRepository.getById(trustRelationshipId);
+    trustRelationship.state = TrustRepository.ENTITY_TRUST_STATE_TYPE.trusted;
+    await this.trustRepository.update(trustRelationship);
   }
 }
