+++ server/models/Wallet.spec.js
@@ -3,6 +3,7 @@ const {expect} = require("chai");
 const jestExpect = require("expect");
 const sinon = require("sinon");
 const WalletRepository = require("../repositories/WalletRepository");
+const TrustRepository = require("../repositories/TrustRepository");
 const WalletService = require("../services/WalletService");
 const HttpError = require("../utils/HttpError");
 
@@ -48,6 +49,62 @@ describe("Wallet", () => {
     WalletRepository.prototype.getByName.restore();
     WalletRepository.prototype.getById.restore();
     Wallet.sha512.restore();
+  });
+
+  it("get trust_relationships", async () => {
+    sinon.stub(TrustRepository.prototype, "get").returns([{a:1}]);
+    sinon.stub(WalletRepository.prototype, "getById").resolves({id:1});
+    const wallet = await walletService.getById(-1);
+    const trust_relationships = await wallet.getTrustRelationships();
+    expect(trust_relationships).lengthOf(1);
+    WalletRepository.prototype.getById.restore();
+    TrustRepository.prototype.get.restore();
+  });
+
+
+  describe("Request trust", () => {
+    let wallet;
+
+    beforeEach(() => {
+      wallet = new Wallet(1);
+    })
+
+    it("request with a wrong type would throw error", async () => {
+      await jestExpect(async () => {
+        await wallet.request("wrongType","test")
+      }).rejects.toThrow("type");
+    });
+
+    it("request with a wrong wallet name would throw error", async () => {
+      await jestExpect(async () => {
+        await wallet.request("send","tes t");
+      }).rejects.toThrow(/name/i);
+    });
+
+    it("request successfully", async () => {
+      sinon.stub(WalletRepository.prototype, "getByName").returns([{id:1}]);
+      sinon.stub(TrustRepository.prototype, "create");
+      await wallet.request("send", "test");
+      WalletRepository.prototype.getByName.restore();
+      TrustRepository.prototype.create.restore();
+    });
+  });
+
+  describe("Accept trust", () => {
+    let wallet;
+
+    beforeEach(() => {
+      wallet = new Wallet(1);
+    })
+
+    it("accept", async () => {
+      sinon.stub(TrustRepository.prototype, "getById").returns([{id:1}]);
+      sinon.stub(TrustRepository.prototype, "update");
+      const trustRelationshipId = 1;
+      await wallet.accept(trustRelationshipId);
+      TrustRepository.prototype.getById.restore();
+      TrustRepository.prototype.update.restore();
+    });
   });
 
 });
